- Add much of pthreads initial implementation - Improve boot menu to be scrollable and offsettable, optimize its redraw - Check for x87 FPU existence - Switch to C++14 - Add proper startup code for efi, store image handle and systab in global - Show which backend is being used for reading boot files to user - Add lots of missed KEEP in linker scripts - Fix CFI for efi startup code - Add several new boot menu options - Fix some types - Fixed broken PXE boot storage layer, can now get file size - Changed boot screen update to be more controlled - Fix some tools being hardcoded instead of discovered in configure - Fix register bitfield generator - Fix incorrect handling of MP spec per-package information - ACPI mmap FADT and DSDT - Add C++ std::atomic etc - Add halt to inline assembly wrappers - Increase max cpus to 512 - Handle relaxed call with dummy prefix in instruction patcher - Fix AMD lfence speculation barrier enable - Add xsave capability getters to cpuid - Schedule __setjmp a bit - Get rid of swapgs boundary annotations - Fix mishandling of some initialization when SMP disabled - Add frame based stacktrace implementation (libgcc does most of the work) - Fix many regressions in gdbstub (but some remain) - Tons of changes to process startup, entry, syscall handling - Tons of changes in memory manager init - Switch everyting to using "at" apis for I/O - Make spinlock optimistically do the cmpxchg then spin on it - Add 32 bit process support! - Implement clone syscall to create user threads - Check in commented out perf test hack - Add _assume to conditionally __builtin_unreachable for you - Added 32 bit nofault cmpxchg - AHCI driver improvements - conditional port e9 debug output - Stop using rdtscp in tracing build hot tracing code - Fix bitfield script to not emit lots of empty lines sometimes - Rough in I/O cancellation (not completed) - IDE driver improvements - Change some locks off mcslock in preparation for EINTR wakeups - Fix several PCI(e) bugs - Many USB improvements - TONS of userspace improvements - Lots more, I got tired of adding more to this commit message

+++ boot/debug.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "types.h"
 
-void debug_out(char const *s, int len);
-void debug_out(char16_t const *s, int len);
+void debug_out(char const *s, ptrdiff_t len);
+void debug_out(char16_t const *s, ptrdiff_t len);
