+++ kernel/lib/engunit.cc
@@ -1,3 +1,3 @@
 #include "engunit.h"
 
-template class engineering_t<uintptr_t>;
+template class engineering_t<uint64_t>;
