+++ libc/src/sys/mman/mremap.cc
@@ -6,8 +6,10 @@
 void *mremap(void *old_address, size_t old_size, size_t new_size,
              int flags, void *new_address)
 {
-    void *result = (void*)syscall5(long(old_address), old_size, new_size,
-                                   flags, long(new_address), SYS_mremap);
+    void *result = (void*)syscall5(uintptr_t(old_address),
+                                   old_size, new_size,
+                                   unsigned(flags), uintptr_t(new_address),
+                                   SYS_mremap);
 
     if ((unsigned long)result > 256)
         return result;
