+++ libc/src/unistd/fchownat.cc
@@ -6,7 +6,8 @@
 
 int fchownat(int dirfd, char const *path, uid_t uid, gid_t gid, int flags)
 {
-    long status = syscall5(dirfd, long(path), uid, gid, flags, SYS_fchownat);
+    long status = syscall5(dirfd, uintptr_t(path),
+                           uid, gid, flags, SYS_fchownat);
 
     if (status >= 0)
         return status;
