+++ libc/src/unistd/lchown.cc
@@ -6,7 +6,7 @@
 
 int lchown(char const *path, uid_t uid, gid_t gid)
 {
-    long status = syscall3(long(path), uid, gid, SYS_lchown);
+    long status = syscall3(uintptr_t(path), uid, gid, SYS_lchown);
 
     if (status >= 0)
         return status;
