+++ libc/src/unistd/readlink.cc
@@ -6,7 +6,7 @@
 
 ssize_t readlink(char const *restrict path, char *restrict buf, size_t sz)
 {
-    long status = syscall3(long(path), long(buf), sz, SYS_readlink);
+    long status = syscall3(uintptr_t(path), uintptr_t(buf), sz, SYS_readlink);
 
     if (status >= 0)
         return status;
